Working on getting URL param
Was able to get the URL param by using let urlString = window.location.href and then split the string into an array so that I could get the param at the very end. 
Currently working on using useParams() to get the urlParam instead. Trying to fix a bug that appears when useParam() is invoked.
Cleaned up index.html.