Navigation and cart functionality implemented
Currently working on updating the cart after the user pressed the 'add # to order' button in the MeatIten Modal. Will add the add-ons option functionality for ordering. 

The user can now go back to the results/Home page of the application by using the side NavBar menu. The side navBar will spring from the left side after the user clicks on the hamburger menu.  More CSS--including the animation when the side navBar closes--will be added.

The modal for the cart icon has been added. The modal will show the current order that the user placed in the cart when the cart has been clicked.  Within the modal, the user can either edit the current order, clear the cart, or proceed to checkout. The modal and its button functionality are still being coded.

Added the Policy, About, and Favorites components for the SideNavBar navigation. Favorites will show all of the restaurants that were 'hearted' by the user.