AddOnItem component built
The AddOnItem component adds an add-on to a user's order when the user presses the check box. This component displays the add-ons to the user in the modal that displays the meat that the user clicked on from either the menu page or from the search bar in the NavBar component.  

Bugs when adding the add-on price to the total price of the order due to how the computer handles numbers since all of the numbers are 'double-precision 64-bit format IEEE 754 values.' For example, when I add 'cheese'  (.99) to a quarter pounder ($1.50) and then take off the cheese, the default price for the quarter pounder then becomes '1.5000000000000002.' Bug will be fixed.