arrays of names and prices of addOns deleted
selectedAddOnsInfoToOrder will store the name and price of the selected addOns of the user as objects in an array. 
The computations for the total price of the addOns are taken from this array. 
The arrays that contained the names of the addOns and the price of the addOns were deleted.